move Button.style into subfolder
